+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Pkg v1.8 Release Notes
 ======================
 
-- New `compat::Bool` kwarg to `Pkg.status` (`--compat` in the REPL mode) to show
+- New `outdated::Bool` kwarg to `Pkg.status` (`--outdated` or `-o` in the REPL mode) to show
   information about packages not at the latest version.
 
 Pkg v1.7 Release Notes